+++ pkg/grpc/api_test.go
@@ -351,6 +351,103 @@ func TestCallRemoteAppWithTracing(t *testing.T) {
 	assert.NotEmpty(t, resp.GetMessage(), "failed to generate trace context with app call")
 }
 
+func TestCallLocalStream(t *testing.T) {
+	t.Run("appchannel is not ready", func(t *testing.T) {
+		port, _ := freeport.GetFreePort()
+
+		fakeAPI := &api{
+			id:         "fakeAPI",
+			appChannel: nil,
+		}
+		server := startInternalServer(port, fakeAPI)
+		defer server.Stop()
+		clientConn := createTestClient(port)
+		defer clientConn.Close()
+
+		client := internalv1pb.NewServiceInvocationClient(clientConn)
+		st, err := client.CallLocalStream(context.Background())
+		assert.NoError(t, err)
+
+		request := invokev1.NewInvokeMethodRequest("method")
+		defer request.Close()
+		err = st.Send(&internalv1pb.InternalInvokeRequestStream{
+			Request: request.Proto(),
+			Payload: &commonv1pb.StreamPayload{Complete: true},
+		})
+		assert.NoError(t, err)
+
+		_, err = st.Recv()
+		assert.Equal(t, codes.Internal, status.Code(err))
+	})
+
+	t.Run("parsing InternalInvokeRequest is failed", func(t *testing.T) {
+		port, _ := freeport.GetFreePort()
+
+		mockAppChannel := new(channelt.MockAppChannel)
+		fakeAPI := &api{
+			id:         "fakeAPI",
+			appChannel: mockAppChannel,
+		}
+		server := startInternalServer(port, fakeAPI)
+		defer server.Stop()
+		clientConn := createTestClient(port)
+		defer clientConn.Close()
+
+		client := internalv1pb.NewServiceInvocationClient(clientConn)
+		st, err := client.CallLocalStream(context.Background())
+		assert.NoError(t, err)
+
+		err = st.Send(&internalv1pb.InternalInvokeRequestStream{
+			Request: &internalv1pb.InternalInvokeRequest{
+				Message: nil,
+			},
+			Payload: &commonv1pb.StreamPayload{Complete: true},
+		})
+		assert.NoError(t, err)
+
+		_, err = st.Recv()
+		assert.Equal(t, codes.InvalidArgument, status.Code(err))
+	})
+
+	t.Run("invokemethod returns error", func(t *testing.T) {
+		port, _ := freeport.GetFreePort()
+
+		mockAppChannel := new(channelt.MockAppChannel)
+		mockAppChannel.On("InvokeMethod", mock.AnythingOfType("*context.cancelCtx"), mock.AnythingOfType("*v1.InvokeMethodRequest")).Return(nil, status.Error(codes.Unknown, "unknown error"))
+		fakeAPI := &api{
+			id:         "fakeAPI",
+			appChannel: mockAppChannel,
+		}
+		server := startInternalServer(port, fakeAPI)
+		defer server.Stop()
+		clientConn := createTestClient(port)
+		defer clientConn.Close()
+
+		client := internalv1pb.NewServiceInvocationClient(clientConn)
+		st, err := client.CallLocalStream(context.Background())
+		assert.NoError(t, err)
+
+		request := invokev1.NewInvokeMethodRequest("method").
+			WithMetadata(map[string][]string{invokev1.DestinationIDHeader: {"foo"}})
+		defer request.Close()
+
+		pd, err := request.ProtoWithData()
+		assert.NoError(t, err)
+
+		err = st.Send(&internalv1pb.InternalInvokeRequestStream{
+			Request: request.Proto(),
+			Payload: &commonv1pb.StreamPayload{
+				Data:     pd.Message.Data,
+				Complete: true,
+			},
+		})
+		assert.NoError(t, err)
+
+		_, err = st.Recv()
+		assert.Equal(t, codes.Internal, status.Code(err))
+	})
+}
+
 func TestCallLocal(t *testing.T) {
 	t.Run("appchannel is not ready", func(t *testing.T) {
 		port, _ := freeport.GetFreePort()
@@ -409,9 +506,10 @@ func TestCallLocal(t *testing.T) {
 		defer clientConn.Close()
 
 		client := internalv1pb.NewServiceInvocationClient(clientConn)
-		request := invokev1.NewInvokeMethodRequest("method").Proto()
+		request := invokev1.NewInvokeMethodRequest("method")
+		defer request.Close()
 
-		_, err := client.CallLocal(context.Background(), request)
+		_, err := client.CallLocal(context.Background(), request.Proto())
 		assert.Equal(t, codes.Internal, status.Code(err))
 	})
 }
